fix: resolve pkg-file-count violations for MCP serve wiring
Move MCP command registration into pkg/cli/mcp and fold PreviewToolResponse
into pkg/mcp/factorysession/tool.go so Build, Lint, and API passes.
